+++ bigtree/_plugins.py
@@ -113,6 +113,17 @@ def register_tree_plugins() -> None:
             "to_pillow": export.tree_to_pillow,
             "to_mermaid": export.tree_to_mermaid,
             "to_vis": export.tree_to_vis,
+            # Iterator methods
+            "preorder_iter": iterators.preorder_iter,
+            "postorder_iter": iterators.postorder_iter,
+            "levelorder_iter": iterators.levelorder_iter,
+            "levelordergroup_iter": iterators.levelordergroup_iter,
+            "zigzag_iter": iterators.zigzag_iter,
+            "zigzaggroup_iter": iterators.zigzaggroup_iter,
+            # Modify methods
+            "shift_nodes": modify.shift_nodes,
+            "copy_nodes": modify.copy_nodes,
+            "shift_and_replace_nodes": modify.shift_and_replace_nodes,
             # Query methods
             "query": query.query_tree,
             # Search methods
@@ -130,17 +141,6 @@ def register_tree_plugins() -> None:
             "find_children": search.find_children,
             "find_child": search.find_child,
             "find_child_by_name": search.find_child_by_name,
-            # Iterator methods
-            "preorder_iter": iterators.preorder_iter,
-            "postorder_iter": iterators.postorder_iter,
-            "levelorder_iter": iterators.levelorder_iter,
-            "levelordergroup_iter": iterators.levelordergroup_iter,
-            "zigzag_iter": iterators.zigzag_iter,
-            "zigzaggroup_iter": iterators.zigzaggroup_iter,
-            # Modify methods
-            "shift_nodes": modify.shift_nodes,
-            "copy_nodes": modify.copy_nodes,
-            "shift_and_replace_nodes": modify.shift_and_replace_nodes,
         }
     )
     Tree.register_plugins(