+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added:
 - Misc: Citation file.
+### Changed:
+- Docs: Clean up and make more succinct.
 
 ## [1.3.1] - 2026-02-14
 ### Added: