+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.4.0] - 2026-03-24
 ### Added:
 - Misc: Citation file.
 - Tree Export: Export to HTML, which also allows tree to be displayed interactively on jupyter notebook.
@@ -907,7 +909,8 @@ ignore null attribute columns.
 - Utility Iterator: Tree traversal methods.
 - Workflow To Do App: Tree use case with to-do list implementation.
 
-[Unreleased]: https://github.com/kayjan/bigtree/compare/1.3.1...HEAD
+[Unreleased]: https://github.com/kayjan/bigtree/compare/1.4.0...HEAD
+[1.4.0]: https://github.com/kayjan/bigtree/compare/1.3.1...1.4.0
 [1.3.1]: https://github.com/kayjan/bigtree/compare/1.3.0...1.3.1
 [1.3.0]: https://github.com/kayjan/bigtree/compare/1.2.0...1.3.0
 [1.2.0]: https://github.com/kayjan/bigtree/compare/1.1.0...1.2.0